backend/sys_GP/requirements.txt
Django==4.2.7
djangorestframework==3.15.0  # ajoute DRF
django-crispy-forms==2.0
django-widget-tweaks==1.5.0
celery==5.3.6
boto3==1.40.3
botocore==1.40.3
requests==2.31.0
PyJWT==2.7.0
cryptography==41.0.7
pillow==11.2.1
